move stories to stories/Link
