+++ packages/react-components/react-link/src/stories/Link/LinkAppearance.stories.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { Link } from '../index';
+import { Link } from '../../index';
 
 export const Appearance = () => (
   <Link appearance="subtle" href="https://www.bing.com">
