+++ packages/react-components/react-link/src/stories/Link/LinkDefault.stories.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { Link, LinkProps } from '../index';
+import { Link, LinkProps } from '../../index';
 
 export const Default = (props: LinkProps & { as?: 'a' }) => (
   <Link href="https://www.bing.com" {...props}>
