+++ packages/react-components/react-link/src/stories/Link/LinkDisabledFocusable.stories.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { Link } from '../index';
+import { Link } from '../../index';
 
 export const DisabledFocusable = () => (
   <Link inline disabled disabledFocusable href="https://www.bing.com">
