+++ packages/react-components/react-link/src/stories/Link/LinkInline.stories.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { Link } from '../index';
+import { Link } from '../../index';
 
 export const Inline = () => (
   <div>
